Add flow control
Currently the FLAC reader sleeps for hard-coded 5ms. This
wakes up the thread more often than it is necessary.
The sleep time could be determined based on the audio
device buffer size. But this has the drawback that the
thread is woken up at an arbritrary time compared to the
audio device thread. Instead in this commit the audio device
thread simply notifies the FLAC thread when it took data
from the AudioBuffer. The audio device thread is realtime
critical so it only `Notify()` the other thread. Thus the
audio device thread stays wait-free. The FLAC reader thread
on the other hand can handle additional jitter given the size
of the AudioBuffer.